requirements.txt
Flask==1.1.2
python-telegram-bot==13.3
waitress==2.0.0